+++ src/components/index.ts
@@ -1,0 +1,16 @@
+export * from './Button';
+export * from './Card';
+export * from './Carousel';
+export * from './CarouselItem';
+export * from './CarouselItem';
+export * from './Divided';
+export * from './Fab';
+export * from './IconButton';
+export * from './IconButton';
+export * from './ProgressIndicator';
+export * from './Slider';
+export * from './Snackbar';
+export * from './Switch';
+export * from './Tab';
+export * from './Tabs';
+export * from './TextField';
