+++ src/effects/custom-scroll/index.ts
@@ -1,3 +1,3 @@
 export * from './custom-scroll.effect';
-export * from './custom-scroll.interface';
 export * from './custom-scroll.style';
+export * from './custom-scroll.interface';
