+++ src/effects/index.ts
@@ -1,2 +1,3 @@
-export * from './smooth-scroll.effect';
+export * from './ripple';
 export * from './custom-scroll';
+export * from './smooth-scroll.effect';
