+++ src/index.ts
@@ -1,15 +1,6 @@
-export * from './components/action/button';
-export * from './components/action/fab';
-export * from './components/action/icon-button';
-export * from './components/communication/progress-indicator';
-export * from './components/communication/snackbar';
-export * from './components/containment/card';
-export * from './components/containment/carousel';
-export * from './components/containment/divider';
-export * from './components/navigation/tabs';
-export * from './components/selection/slider';
-export * from './components/selection/switch';
+export * from './components';
 export * from './effects';
 export * from './icon';
-export * from './text-inputs';
+export * from './interfaces';
+export * from './styles';
 export * from './utils';
