+++ src/interfaces/index.ts
@@ -1,0 +1,14 @@
+export * from './button.interface';
+export * from './card.interface';
+export * from './carousel-item.interface';
+export * from './carousel.interface';
+export * from './divider.interface';
+export * from './fab.interface';
+export * from './icon-button.interface';
+export * from './progress-indicator.interface';
+export * from './slider.interface';
+export * from './snackbar.interface';
+export * from './switch.interface';
+export * from './tab.interface';
+export * from './tabs.interface';
+export * from './text-field.interface';
