+++ src/styles/carousel.style.ts
