+++ src/utils/index.ts
@@ -1,2 +1,7 @@
 export * from './component-helper';
+export * from './component';
+export * from './component';
+export * from './component';
+export * from './component';
+export * from './component';
 export * from './styles';
