+++ src/utils/styles/index.ts
@@ -1,2 +1,3 @@
 export * from './classnames';
+export * from './classnames';
 export * from './get-classname';
